pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "databricks-sql-connector>=4.2.5",
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
